+++ src/ext/dispatch.c
@@ -136,16 +136,18 @@ static void execute_fcall(ddtrace_dispatch_t *dispatch, zend_execute_data *execu
     if (fci.params) {
         efree(fci.params);
     }
+#else
+    if (fci.params) {
+        zend_fcall_info_args_clear(&fci, 0);
+    }
 #endif
 
 _exit_cleanup:
-#if PHP_VERSION_ID < 70000
     if (this) {
         Z_DELREF_P(this);
     }
 
-#endif
-    zval_dtor(&closure);
+    Z_DELREF(closure);
 }
 
 static int is_anonymous_closure(zend_function *fbc, const char *function_name, uint32_t *function_name_length_p) {
@@ -204,7 +206,7 @@ static zend_always_inline zend_bool wrap_and_run(zend_execute_data *execute_data
         common_scope_length = ZSTR_LEN(fbc->common.scope->name);
 #endif
     }
-    DD_PRINTF("Loaded object id: %0lx", object);
+    DD_PRINTF("Loaded object id: %p", (void *)object);
 
     ddtrace_dispatch_t *dispatch;
 
@@ -283,8 +285,13 @@ static zend_always_inline zend_bool wrap_and_run(zend_execute_data *execute_data
         execute_fcall(dispatch, execute_data, &return_value TSRMLS_CC);
 
         if (return_value != NULL) {
-            EX_TMP_VAR(execute_data, opline->result.var)->var.ptr = return_value;
+            if (RETURN_VALUE_USED(opline)) {
+                EX_TMP_VAR(execute_data, opline->result.var)->var.ptr = return_value;
+            } else {
+                zval_ptr_dtor(&return_value);
+            }
         }
+
 #else
         zval rv;
         INIT_ZVAL(rv);