+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file - [read more
 
 ## [UNRELEASED]
 
+### Fixed
+- Memory leaks in `this` object and return value handling in PHP 5.6 and 7.x
+
 ## [0.11.0]
 
 **WARNING: THIS IS A BREAKING CHANGE RELEASE**